Update CI skips/fails (#1582)
* Update CI skips/fails

- clean up skipifs that should not be relevant anymore since said
  APIs have been landed for a while

* cuda isn't available in CI, use aot_eager

* Consolidate tests using distributed processgroup

- move existing test_misc test into test_distributed
- avoid creating/destroying multiple pg's per process as it seems
  to hang sometimes (unclear if this indicates a pg bug but WAR
  for now)

* lint
